Supported x86_64.
git-svn-id: https://cairo-xemmai.googlecode.com/svn/trunk@25 5d80cac4-d776-11de-8a94-ab732a2dc7d9

+++ cairo/xemmai/cairo.cc
@@ -185,7 +185,7 @@ namespace xemmai
 
 void t_type_of<cairo_status_t>::f_define(t_extension* a_extension)
 {
-	t_define<cairo_status_t, int>(a_extension, L"Status")
+	t_define<cairo_status_t, ptrdiff_t>(a_extension, L"Status")
 		(L"SUCCESS", CAIRO_STATUS_SUCCESS)
 		(L"NO_MEMORY", CAIRO_STATUS_NO_MEMORY)
 		(L"INVALID_RESTORE", CAIRO_STATUS_INVALID_RESTORE)
